Adicionar as pastas

+++ src/main/java/com/example/literalura/model/DadosLivros.java
@@ -6,8 +6,18 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.List;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record DadosLivros(@JsonAlias("title") String titulo,
-                          @JsonAlias("authors") List<DadosAutores> autor,
-                          @JsonAlias("languages") List<String> idioma,
-                          @JsonAlias("download_count") Double numeroDownloads) {
+public record DadosLivros(
+
+        @JsonAlias("title")
+        String titulo,
+
+        @JsonAlias("authors")
+        List<DadosAutores> autor,
+
+        @JsonAlias("languages")
+        List<String> idioma,
+
+        @JsonAlias("download_count")
+        Integer numeroDownloads
+) {
 }
